requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.1
requests>=2.32.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
